separate Params into own file
